server/server/requirements.txt
asgiref==3.8.1
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.42.2
dj-database-url==3.0.0
django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
pyjwt==2.9.0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
